+++ src/components/Header/Header.css
@@ -1,4 +1,4 @@
-#headerShared {
+.headerShared {
   position: sticky;
   display: flex;
   justify-content: space-between;
@@ -8,7 +8,7 @@
   background-color: var(--Linkbrary-bg, #f0f6ff);
 }
 
-#headerFolder {
+.headerFolder {
   position: static;
   display: flex;
   width: 100%;
